requirements.txt
fastapi
uvicorn
pydantic
openai

sqlalchemy>=2.0
asyncpg
python-dotenv

